+++ tests/help.commandDescription.test.js
@@ -1,0 +1,20 @@
+const commander = require('../');
+
+// These are tests of the Help class, not of the Command help.
+// There is some overlap with the higher level Command tests (which predate Help).
+
+describe('commandDescription', () => {
+  test('when program has no description then empty string', () => {
+    const program = new commander.Command();
+    const helper = new commander.Help();
+    expect(helper.commandDescription(program)).toEqual('');
+  });
+
+  test('when program has description then return description', () => {
+    const description = 'description';
+    const program = new commander.Command();
+    program.description(description);
+    const helper = new commander.Help();
+    expect(helper.commandDescription(program)).toEqual(description);
+  });
+});
