+++ tests/help.sortCommands.test.js
@@ -1,0 +1,29 @@
+const commander = require('../');
+
+// These are tests of the Help class, not of the Command help.
+// There is some overlap with the higher level Command tests (which predate Help).
+
+describe('sortCommands', () => {
+  test('when unsorted then commands in order added', () => {
+    const program = new commander.Command();
+    program
+      .command('ccc', 'desc')
+      .command('aaa', 'desc')
+      .command('bbb', 'desc');
+    const helper = program.createHelp();
+    const visibleCommandNames = helper.visibleCommands(program).map(cmd => cmd.name());
+    expect(visibleCommandNames).toEqual(['ccc', 'aaa', 'bbb', 'help']);
+  });
+
+  test('when sortCommands:true then commands sorted', () => {
+    const program = new commander.Command();
+    program
+      .configureHelp({ sortCommands: true })
+      .command('ccc', 'desc')
+      .command('aaa', 'desc')
+      .command('bbb', 'desc');
+    const helper = program.createHelp();
+    const visibleCommandNames = helper.visibleCommands(program).map(cmd => cmd.name());
+    expect(visibleCommandNames).toEqual(['aaa', 'bbb', 'ccc', 'help']);
+  });
+});
